feat/refactor: SemiAuto / Burst mode toggle on G18
Old code would not support Burst because the top level Player was polling the squeeze method and then calling the fire() function if squeeze returned true. This meant that burst would fundamentally not work correctly, since if you let go of the trigger mid-burst then it would stop. Also, things are just way easier when you use a Timer for a trigger release delay anyway, for semi auto too. Full auto is always the easiest to implement anyway, but I have no full auto weapon to implement it for yet.
Also moved away from a polymorphic subclass approach to simple fire modes, and use a switch statement in the fire policy. Otherwise, I have to create a wrapper on the fire policy to actually HOLD the true fire policy/strategy, which I don't want to do for something relatively small like this. Maybe later, if we truly need flexibility for tons more fire modes.